Search for deps in all sketch dir source files
Automatic dependency search is now applied to all source files in
the sketch directory when the LIBS variable is empty.
To avoid unintentional building and flashing the demo application
whenever a sketch could not be found, this now has to be activated
explicitely by using DEMO=1 on the command line
Some adjustments for cygwin.
Corrected parameters for esptool in restore_flash.