Move sketch prebuild inside conditional USE_PREBUILD
Execution of all actions coming from "prebuild" sections in the Arduino
platform.txt are now made conditional to the variable USE_PREBUILD.
The main reason for this is that the signing fuctionality in
"tools/signing.py" generates a dummy header file every
time the application is linked, even if there is no public signing key
available. This is done although such a dummy file is present in the repo.
This in turn trigger unneccesary rebuilds of Updater.cpp and other files
every time.
If you are using the signing functionality however and have a public key you
must set USE_PREBUILD