Enable inclusion of all source files in LIBS
If the implementation source file is located in another directory
than the corresponding header file, this file may not be included
in the build by the automatic search of used #includes.
A new variable named EXPAND_LIBS can be set to force include all the
source files found in directories specified via LIBS in the build.